Game starts up again, but crashes immediately on clicks
