+++ minegauler/core/__init__.py
@@ -1,18 +1,6 @@
 # April 2018, Lewis Gaul
 
-__all__ = ("GAME_MODE_IMPL",)
+__all__ = "UberController"
 
-from typing import Any, Mapping
-
-from ..shared.types import GameMode
-from . import board, game, minefield, regular, split_cells
-
-
-# TODO: Re-add imports when fixed
-# from . import api, board, engine, game, minefield, regular, split_cells
-
-
-GAME_MODE_IMPL: Mapping[GameMode, Any] = {
-    GameMode.REGULAR: regular,
-    GameMode.SPLIT_CELL: split_cells,
-}
+from . import api, board, engine, game, minefield, regular, split_cells
+from .engine import UberController
