+++ minegauler/core/regular/__init__.py
@@ -1,1 +1,9 @@
 # October 2021, Lewis Gaul
+
+__all__ = ("Coord", "Minefield", "Board", "Game", "Controller")
+
+from .board import Board
+from .controller import Controller
+from .game import Game
+from .minefield import Minefield
+from .types import Coord
