+++ minegauler/core/regular/controller.py
@@ -5,20 +5,12 @@ __all__ = ("Controller",)
 import json
 import logging
 import os.path
-import sys
-from typing import Dict, Optional
 
+import minegauler.shared.utils
 
-if sys.version_info >= (3, 8):
-    from typing import Literal
-else:
-    from typing_extensions import Literal
-
-from attr import attr
-
-from ...shared import GameOptsStruct
+from ...shared import GameOptsStruct, utils
 from ...shared.types import CellContents, GameMode, GameState, PathLike
-from .. import api
+from .. import api, game
 from ..controller import ControllerBase
 from .board import Board
 from .game import Game
@@ -27,11 +19,6 @@ from .types import Coord
 
 
 logger = logging.getLogger(__name__)
-
-
-# TODO:
-#  This whole module needs effectively rewriting from scratch, hasn't worked
-#  since stuff was moved around.
 
 
 def _save_minefield(mf: Minefield, file: PathLike) -> None:
@@ -51,90 +38,13 @@ def _save_minefield(mf: Minefield, file: PathLike) -> None:
         json.dump(mf.to_json(), f)
 
 
-@attr.attrs(auto_attribs=True, kw_only=True)
-class _SharedInfo:
-    """
-    Information to pass to frontends.
-
-    Elements:
-    cell_updates
-        Dictionary of updates to cells, mapping the coordinate to the new
-        contents of the cell.
-    game_state
-        The state of the game.
-    mines_remaining
-        The number of mines remaining to be found, given by
-        [total mines] - [number of flags].
-        Can be negative if there are too many flags.
-    lives_remaining
-        The number of lives remaining.
-    """
-
-    cell_updates: Optional[Dict[Coord, CellContents]] = None
-    game_state: GameState = GameState.READY
-    mines_remaining: int = 0
-    lives_remaining: int = 0
-
-
-class Controller(ControllerBase[Literal[GameMode.REGULAR]]):
+class Controller(ControllerBase):
     """Controller for a regular minesweeper game."""
 
     mode = GameMode.REGULAR
+    minefield_cls = Minefield
     board_cls = Board
     game_cls = Game
-
-    def __init__(
-        self, opts: GameOptsStruct, *, notif: api.AbstractListener,
-    ):
-        """
-        :param opts:
-            Game options.
-        :param notif:
-            A notifier defining callbacks.
-        """
-        super().__init__(opts)
-        # Use a reference to the given opts rather than a copy.
-        self._opts = opts
-        self._notif = notif
-        self._game = self.game_cls(
-            x_size=self._opts.x_size,
-            y_size=self._opts.y_size,
-            mines=self._opts.mines,
-            per_cell=self._opts.per_cell,
-            lives=self._opts.lives,
-            first_success=self._opts.first_success,
-        )
-        self._last_update = _SharedInfo()
-        self._send_updates()
-        self._notif.set_mines(self._opts.mines)
-
-    @property
-    def board(self) -> Board:
-        return self._game.board
-
-    def get_game_info(self) -> api.GameInfo:
-        """Get info about the current game."""
-        ret = api.GameInfo(
-            game_state=self._game.state,
-            x_size=self._game.x_size,
-            y_size=self._game.y_size,
-            mines=self._game.mines,
-            difficulty=self._game.difficulty,
-            per_cell=self._game.per_cell,
-            first_success=self._game.first_success,
-            minefield_known=self._game.minefield_known,
-        )
-        if self._game.state.started():
-            ret.started_info = api.GameInfo.StartedInfo(
-                start_time=self._game.start_time,
-                elapsed=self._game.get_elapsed(),
-                bbbv=self._game.mf.bbbv,
-                rem_bbbv=self._game.get_rem_3bv(),
-                bbbvps=self._game.get_3bvps(),
-                prop_complete=self._game.get_prop_complete(),
-                prop_flagging=self._game.get_flag_proportion(),
-            )
-        return ret
 
     # --------------------------------------------------------------------------
     # Methods triggered by user interaction
@@ -323,25 +233,142 @@ class Controller(ControllerBase[Literal[GameMode.REGULAR]]):
         self._notif.set_mines(self._opts.mines)
         self._send_updates()
 
-    def _send_updates(
-        self, cells_updated: Optional[Dict[Coord, CellContents]] = None
-    ) -> None:
-        """Send updates to registered listeners."""
-        update = _SharedInfo(
-            cell_updates=cells_updated,
-            mines_remaining=self._game.mines_remaining,
-            lives_remaining=self._game.lives_remaining,
-            game_state=self._game.state,
+
+class CreateController(ControllerBase):
+    """A controller for creating boards."""
+
+    def __init__(self, opts: GameOptsStruct, *, notif: api.AbstractListener):
+        super().__init__(opts)
+        # Use a reference to the given opts rather than a copy.
+        self._opts = opts
+        self._notif = notif
+        self._board = Board(self._opts.x_size, self._opts.y_size)
+        self._flags: int = 0
+        self._notif.set_mines(self._flags)
+
+    @property
+    def board(self) -> Board:
+        return self._board
+
+    def get_game_info(self) -> api.GameInfo:
+        return api.GameInfo(
+            game_state=GameState.ACTIVE,
+            x_size=self._opts.x_size,
+            y_size=self._opts.y_size,
+            mines=self._flags,
+            difficulty=minegauler.shared.utils.difficulty_from_values(
+                GameMode.REGULAR, self._opts.x_size, self._opts.x_size, self._flags
+            ),
+            per_cell=self._opts.per_cell,
+            first_success=self._opts.first_success,
+            minefield_known=True,
         )
 
-        # Send updates to registered listeners.
-        if update.cell_updates:
-            self._notif.update_cells(update.cell_updates)
-        if update.mines_remaining != self._last_update.mines_remaining:
-            self._notif.update_mines_remaining(update.mines_remaining)
-        # if update.lives_remaining != self._last_update.lives_remaining:
-        #     self._notif.update_lives_remaining(update.lives_remaining)
-        if update.game_state is not self._last_update.game_state:
-            self._notif.update_game_state(update.game_state)
+    def new_game(self) -> None:
+        """See AbstractController."""
+        super().new_game()
+        self._board = self.board_cls(self._opts.x_size, self._opts.y_size)
+        self._flags = 0
+        self._notif.reset()
 
-        self._last_update = update
+    def restart_game(self) -> None:
+        super().restart_game()
+        self.new_game()
+
+    def select_cell(self, coord: Coord) -> None:
+        super().select_cell(coord)
+        cell = self._board[coord]
+        if cell is CellContents.Unclicked:
+            self._board[coord] = CellContents.Num(0)
+        elif isinstance(cell, CellContents.Num):
+            self._board[coord] += 1
+        else:
+            return
+        self._notif.update_cells({coord: self._board[coord]})
+
+    def flag_cell(self, coord: Coord, *, flag_only: bool = False) -> None:
+        super().flag_cell(coord)
+        cell = self._board[coord]
+
+        if cell is CellContents.Unclicked:
+            self._board[coord] = CellContents.Mine(1)
+            self._flags += 1
+        elif isinstance(cell, CellContents.Mine):
+            if cell.num == self._opts.per_cell:
+                if flag_only:
+                    return
+                self._board[coord] = CellContents.Unclicked
+                self._flags -= self._opts.per_cell
+            else:
+                self._board[coord] += 1
+                self._flags += 1
+        elif isinstance(cell, CellContents.Num):
+            self.board[coord] = CellContents.Unclicked
+        else:
+            return
+        self._notif.update_cells({coord: self._board[coord]})
+        self._notif.update_mines_remaining(self._flags)
+
+    def chord_on_cell(self, coord: Coord) -> None:
+        super().chord_on_cell(coord)
+
+    def remove_cell_flags(self, coord: Coord) -> None:
+        super().remove_cell_flags(coord)
+        cell = self._board[coord]
+        if not isinstance(cell, CellContents.Mine):
+            return
+        self._board[coord] = CellContents.Unclicked
+        self._flags -= cell.num
+        self._notif.update_cells({coord: self._board[coord]})
+        self._notif.update_mines_remaining(self._flags)
+
+    def resize_board(self, x_size: int, y_size: int, mines: int) -> None:
+        """Resize the board. The number of mines is ignored."""
+        super().resize_board(x_size=x_size, y_size=y_size, mines=mines)
+        if x_size == self._opts.x_size and y_size == self._opts.y_size:
+            logger.info(
+                "No resize required as the parameters are unchanged, starting a new game"
+            )
+            self.new_game()
+            return
+
+        logger.info(
+            "Resizing board from %sx%s to %sx%s",
+            self._opts.x_size,
+            self._opts.y_size,
+            x_size,
+            y_size,
+        )
+        self._opts.x_size = x_size
+        self._opts.y_size = y_size
+        self._notif.resize_minefield(x_size, y_size)
+        self.new_game()
+
+    def set_first_success(self, value: bool) -> None:
+        super().set_first_success(value)
+        self._opts.first_success = value
+
+    def set_per_cell(self, value: int) -> None:
+        super().set_per_cell(value)
+        self._opts.per_cell = value
+
+    def save_current_minefield(self, file: PathLike) -> None:
+        """
+        Save the current created minefield to file.
+
+        :param file:
+            The location of the file to save to. Should have the extension
+            ".mgb".
+        :raises OSError:
+            If saving to file fails.
+        """
+        super().save_current_minefield(file)
+        mines_grid = utils.Grid(self._opts.x_size, self._opts.y_size)
+        for c in self._board.all_coords:
+            if type(self._board[c]) is CellContents.Mine:
+                mines_grid[c] = self._board[c].num
+        mf = Minefield.from_grid(mines_grid, per_cell=self._opts.per_cell)
+        _save_minefield(mf, file)
+
+    def load_minefield(self, file: PathLike) -> None:
+        return NotImplemented
