+++ minegauler/frontend/minefield/base.py
@@ -10,29 +10,25 @@ Exports
 
 """
 
-__all__ = ("MinefieldWidget", "SplitCellMinefieldWidget")
+__all__ = ("MinefieldWidgetBase",)
 
-import enum
+import abc
 import functools
 import logging
 import time
-from typing import Any, Callable, Dict, Iterable, List, Mapping, Optional, Set, Tuple
+from typing import Callable, Dict, Iterable, List, Mapping, Optional, Set, Type
 
 from PyQt5.QtCore import QSize, Qt, pyqtSignal
 from PyQt5.QtGui import QImage, QMouseEvent, QPainter, QPixmap
 from PyQt5.QtWidgets import QGraphicsScene, QGraphicsView, QSizePolicy, QWidget
 
-from ..core import Board, api
-from ..core.board import RegularBoard, SplitCellBoard
-from ..shared.types import (
-    CellContents,
-    CellImageType,
-    Coord_T,
-    RegularCoord,
-    SplitCellCoord,
-)
-from .state import State
-from .utils import IMG_DIR, CellUpdate_T, MouseMove
+from minegauler.core import api
+from minegauler.core.board import BoardBase
+from minegauler.core.controller import ControllerBase
+from minegauler.shared.types import CellContents, CellImageType, Coord_T
+
+from ..state import State
+from ..utils import IMG_DIR, CellUpdate_T, MouseMove
 
 
 logger = logging.getLogger(__name__)
@@ -151,17 +147,19 @@ def _filter_left_and_right(mouse_event_func: Callable):
     return wrapper
 
 
-class MinefieldWidget(QGraphicsView):
+class MinefieldWidgetBase(QGraphicsView):
     """
     The minefield widget.
     """
+
+    ctrlr_cls: Type[ControllerBase]
 
     at_risk_signal = pyqtSignal()
     no_risk_signal = pyqtSignal()
     size_changed = pyqtSignal()
 
     def __init__(
-        self, parent: Optional[QWidget], ctrlr: api.AbstractController, state: State,
+        self, parent: Optional[QWidget], ctrlr: ControllerBase, state: State,
     ):
         super().__init__(parent)
         logger.info("Initialising minefield widget")
@@ -194,7 +192,7 @@ class MinefieldWidget(QGraphicsView):
         self.reset()
 
     @property
-    def _board(self) -> Board:
+    def _board(self) -> BoardBase:
         return self._ctrlr.board
 
     @property
@@ -544,7 +542,7 @@ class MinefieldWidget(QGraphicsView):
         self._mouse_events = []
         self._first_click_time = None
 
-    def update_cells(self, cell_updates: Mapping[RegularCoord, CellContents]) -> None:
+    def update_cells(self, cell_updates: Mapping[Coord_T, CellContents]) -> None:
         """
         Called to indicate some cells have changed state.
 
@@ -575,107 +573,3 @@ class MinefieldWidget(QGraphicsView):
 
     def get_mouse_events(self) -> List[CellUpdate_T]:
         return self._mouse_events
-
-
-RightClickAction = enum.Enum("RightClickAction", ["FLAG", "UNFLAG", "SPLIT"])
-
-
-class SplitCellMinefieldWidget(MinefieldWidget):
-    def __init__(self, *args, **kwargs):
-        self._large_cell_images: Dict[CellContents, QPixmap] = {}
-        super().__init__(*args, **kwargs)
-
-        self._right_click_action: Optional[RightClickAction] = None
-        del self._unflag_on_right_drag
-
-    @property
-    def _board(self) -> SplitCellBoard:
-        return self._ctrlr.board
-
-    # --------------------------------------------------------------------------
-    # Mouse click handlers
-    # --------------------------------------------------------------------------
-    def right_button_down(self, coord: SplitCellCoord) -> None:
-        """
-        Right mouse button was pressed. Change display and call callback
-        functions as appropriate.
-        """
-        if coord.is_split:
-            self._ctrlr.flag_cell(coord)
-            if self._board[coord] is CellContents.Unclicked:
-                self._right_click_action = RightClickAction.UNFLAG
-            else:
-                self._right_click_action = RightClickAction.FLAG
-        else:
-            self._ctrlr.split_cell(coord)
-            if self._board.get_coord_at(coord.x, coord.y).is_split:
-                self._right_click_action = RightClickAction.SPLIT
-
-    def right_button_move(self, coord: Optional[SplitCellCoord]) -> None:
-        """
-        Right mouse button was moved. Change display as appropriate.
-        """
-        if self._state.drag_select and coord is not None:
-            if self._right_click_action is RightClickAction.SPLIT:
-                self._ctrlr.split_cell(coord)
-            elif self._right_click_action is RightClickAction.UNFLAG:
-                self._ctrlr.remove_cell_flags(coord)
-            elif self._right_click_action is RightClickAction.FLAG:
-                self._ctrlr.flag_cell(coord, flag_only=True)
-            else:
-                # The right-button down had no effect, so treat the move like
-                # a first click.
-                self.right_button_down(coord)
-
-    # --------------------------------------------------------------------------
-    # Other public methods
-    # --------------------------------------------------------------------------
-    def reset(self) -> None:
-        if isinstance(self._board, RegularBoard):
-            # Wait for the controller to be switched before resetting.
-            return
-        super().reset()
-        self._right_click_action = None
-
-    def update_cells(self, cell_updates: Mapping[SplitCellCoord, Any]) -> None:
-        """
-        Called to indicate some cells have changed state.
-
-        :param cell_updates:
-            A mapping of cell coordinates to their new state.
-        """
-        for c, v in cell_updates.items():
-            if c.is_split:
-                self._set_cell_image(c, v)
-            else:
-                self._set_large_cell_image(c, v)
-
-    def all_buttons_release(self) -> None:
-        super().all_buttons_release()
-        self._right_click_action = None
-
-    # --------------------------------------------------------------------------
-    # Helpers
-    # --------------------------------------------------------------------------
-    def _set_large_cell_image(self, coord: SplitCellCoord, state: CellContents) -> None:
-        if state not in self._large_cell_images:
-            logger.error("Missing cell image for state: %s", state)
-            return
-        b = self._scene.addPixmap(self._large_cell_images[state])
-        b.setPos(coord.x * self.btn_size, coord.y * self.btn_size)
-
-    def _redraw_cells(self):
-        self._scene.clear()
-        for coord in self._board.all_coords:
-            if coord.is_split:
-                self._set_cell_image(coord, self._board[coord])
-            else:
-                self._set_large_cell_image(coord, self._board[coord])
-
-    def _update_cell_images(self, img_type: CellImageType = CellImageType.ALL) -> None:
-        _update_cell_images(
-            self._cell_images, self.btn_size, self._state.styles, img_type
-        )
-        _update_cell_images(
-            self._large_cell_images, self.btn_size * 2, self._state.styles, img_type
-        )
